% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_mapquest}
\alias{geocode_mapquest}
\title{Interface with MapQuest API to geocode addresses}
\usage{
geocode_mapquest(addresses, key)
}
\arguments{
\item{addresses}{A string vector containing addresses, locations, or
points of interest to geocode.}

\item{key}{An individual API authentication key. MapQuest provides
15,000 free monthly transactions with a non-gated account. An
account can be created here:
https://developer.mapquest.com/plan_purchase/steps/business_edition/business_edition_free/register}
}
\value{
A dataframe will be returned with the given addresses and their
corresponding latitudes and longitudes. These are attributed to the
closest street location.
}
\description{
\code{geocode_mapquest} creates and executes calls to MapQuest
batch geocoding API via HTTP GET.
}
\examples{
geocode_mapquest(schools, key = Sys.getenv("MAPQUEST_API_KEY"))


}
