% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_mapquest_reverse}
\alias{geocode_mapquest_reverse}
\title{Interface with MapQuest API to reverse geocode addresses}
\usage{
geocode_mapquest_reverse(latlon, key)
}
\arguments{
\item{latlon}{A dataframe including lat lng pairs to reverse geocode.}

\item{key}{An individual API authentication key. MapQuest provides
15,000 free monthly transactions with a non-gated account. An
account can be created here:
https://developer.mapquest.com/plan_purchase/steps/business_edition/business_edition_free/register}
}
\value{
A dataframe will be returned with the given latitudes and
longitudes as well as their corresponding addresses.
}
\description{
\code{geocode_mapquest_reverse} creates and executes calls to MapQuest
reverse geocoding API via HTTP GET.
}
\examples{
geocode_mapquest_reverse(schools, key = Sys.getenv("MAPQUEST_API_KEY"))


}
